refactor: split province sea zone overlay into scoped parts (#3878)
Move province/sea content assembly and tile-section builders into dedicated
part files so each overlay submodule stays under the panel size target without
behavior changes.

Refs #3878
